# swagger.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","summary":"Register User","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/sso-login":{"post":{"operationId":"AuthController_ssoLogin","summary":"SSO Login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/admin":{"post":{"operationId":"AuthController_getAdminData","summary":"TEST","parameters":[],"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/auth/author":{"post":{"operationId":"AuthController_getAuthorData","summary":"TEST","parameters":[],"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/user/me":{"get":{"operationId":"UserController_getCurrentUser","summary":"Get current user","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/user/all":{"get":{"operationId":"UserController_getAllUsers","summary":"Get all users with pagination","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"searchFields","required":false,"in":"query","description":"Fields to search in","schema":{"type":"array","items":{"type":"string"}}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"enum":["asc","desc"],"type":"string"}}],"responses":{"200":{"description":"Returns paginated list of users"}},"tags":["Users"]}},"/user/{id}":{"patch":{"operationId":"UserController_updateUser","summary":"Update user by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"get":{"operationId":"UserController_getUserById","summary":"Get user by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/health":{"get":{"operationId":"HealthController_check","summary":"Health Check","parameters":[],"responses":{"200":{"description":"Health Check run correctly."},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health Check"]}},"/approvals":{"post":{"operationId":"ApprovalsController_createApproval","summary":"Create approval","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApprovalDto"}}}},"responses":{"201":{"description":""}},"tags":["Approvals"]}},"/approvals/all":{"get":{"operationId":"ApprovalsController_getAllUsers","summary":"Get all approvals with pagination","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"searchFields","required":false,"in":"query","description":"Fields to search in","schema":{"type":"array","items":{"type":"string"}}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"enum":["asc","desc"],"type":"string"}}],"responses":{"200":{"description":"Returns paginated list of approvals"}},"tags":["Approvals"]}},"/approvals/{id}":{"patch":{"operationId":"ApprovalsController_updateApproval","summary":"Update approval","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateApprovalDto"}}}},"responses":{"200":{"description":""}},"tags":["Approvals"]}},"/sop/all":{"get":{"operationId":"SopController_getAllSops","summary":"Get all SOPs","description":"Retrieves a paginated list of all SOPs. Only accessible to ADMIN and AUTHOR roles.","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"searchFields","required":false,"in":"query","description":"Fields to search in","schema":{"type":"array","items":{"type":"string"}}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"enum":["asc","desc"],"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved the list of SOPs."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"tags":["SOPs"],"security":[{"bearer":[]}]}},"/sop/{id}":{"get":{"operationId":"SopController_getSopById","summary":"Get SOP by ID","description":"Retrieves a specific SOP by its ID. Only accessible to ADMIN and AUTHOR roles.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the SOP details with its content."},"401":{"description":"Unauthorized"},"404":{"description":"SOP not found"}},"tags":["SOPs"],"security":[{"bearer":[]}]},"patch":{"operationId":"SopController_updateSop","summary":"Update SOP by ID","description":"Updates the details of a specific SOP by its ID. Only accessible to ADMIN and AUTHOR roles.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"SOP updated successfully."},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"SOP not found"}},"tags":["SOPs"],"security":[{"bearer":[]}]}},"/sop":{"post":{"operationId":"SopController_createSop","summary":"Create a new SOP","description":"Creates a new SOP with the provided data. Only accessible to ADMIN and AUTHOR roles.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSopDto"}}}},"responses":{"201":{"description":"SOP created successfully."},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"tags":["SOPs"],"security":[{"bearer":[]}]}}},"info":{"title":"SOPWISE","description":"API documentation for the SOPWISE project","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"test@test.com"},"name":{"type":"string","example":"John Doe"},"password":{"type":"string","example":"password"},"role":{"type":"string","example":"ASSISTANT"},"provider":{"type":"string","example":"sopwise","default":"sopwise"}},"required":["email","name","password","role","provider"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"test@test.com"},"password":{"type":"string","example":"password"}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"new_email@test.com"},"name":{"type":"string","example":"New Name"},"password":{"type":"string","example":"password"},"role":{"type":"string","example":"ASSISTANT"},"provider":{"type":"string","example":"sopwise","default":"sopwise"},"metaData":{"type":"string","example":"{\"key\": \"value\"}"}}},"CreateApprovalDto":{"type":"object","properties":{"description":{"type":"string","description":"Description of the approval request","example":"Request for project access"},"status":{"type":"string","description":"Status of the approval","enum":["PENDING","APPROVED","REJECTED"],"default":"PENDING"},"approvedBy":{"type":"string","description":"ID of the user approving the request","example":"123e4567-e89b-12d3-a456-426614174000"},"allowedRole":{"type":"array","description":"Roles allowed for this approval","items":{"type":"string","enum":["ADMIN","AUTHOR","ASSISTANT"]}}},"required":["description"]},"UpdateApprovalDto":{"type":"object","properties":{"description":{"type":"string","description":"Description of the approval request","example":"Request for project access"},"status":{"type":"string","description":"Status of the approval","enum":["PENDING","APPROVED","REJECTED"],"default":"PENDING"},"approvedBy":{"type":"string","description":"ID of the user approving the request","example":"123e4567-e89b-12d3-a456-426614174000"},"allowedRole":{"type":"array","description":"Roles allowed for this approval","items":{"type":"string","enum":["ADMIN","AUTHOR","ASSISTANT"]}}}},"CreateSopDto":{"type":"object","properties":{}}}}}